+++ src/test/java/model/TestCategory.java
@@ -34,7 +34,7 @@ public class TestCategory {
 	
 	@Test
 	public void testGetId() {
-		assertEquals(1, tested.getId());
+		assertEquals(1, tested.getId().intValue());
 	}
 	@Test(expected=IllegalArgumentException.class)
 	public void testSetId_isInvalid() {
