+++ src/test/java/model/TestTargetTransaction.java
@@ -32,7 +32,7 @@ public class TestTargetTransaction {
 	public void testGetId() {
 		TargetTransaction tested = setTarget();
 		tested.setId(1);
-		assertEquals(1, tested.getId());
+		assertEquals(1, tested.getId().intValue());
 	}
 	@Test(expected=IllegalArgumentException.class)
 	public void testSetId_isInvalid() {
